requirements.txt
django
django-filter
django-allauth
python-dateutil
markdown
django-resized
python-decouple
django-ckeditor
django-crispy-forms
pillow
six~=1.16.0
Faker
environ
django-phonenumber-field
django-phonenumbers
django-environ
DateTime~=5.0
djangorestframework
exarth-rest-auth
drf_yasg
openai
